requirements.txt
ultralytics
opencv-python
python-telegram-bot
python-dotenv
loguru
cmake
dlib
face_recognition
customtkinter
Pillow

# RTMDet & Cache Requirements
mmdet>=3.0.0
mmcv>=2.0.0,<2.2.0
torch>=2.0.0
torchvision>=0.15.0
redis>=4.5.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
numpy>=1.24.0,<2.0.0
pytest>=7.3.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Event Streaming & Monitoring
kafka-python>=2.0.2
prometheus-client>=0.17.0
psutil>=5.9.0

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
